refactor: divid classes into two files + cosmetic fixes
