+++ game/main.py
@@ -1,14 +1,16 @@
 import asyncio
+import random
 import sys
 
 import pygame
-import random
 
-from player import Player, Brick
+from brick import Brick
+from player import Player
 
 
 FPS = 30
 SCREEN_SIZE = (1200, 800)
+NUMBER_OF_BRICKS = random.randint(1, 10)
 
 
 async def main():
@@ -25,7 +27,7 @@ async def main():
     all_sprites = pygame.sprite.RenderPlain((player, ))
 
     # Создаем кирпичи
-    bricks = [Brick() for _ in range(random.randint(1, 10))]
+    bricks = [Brick() for _ in range(NUMBER_OF_BRICKS)]
     all_sprites.add(bricks)
 
     while True:
